Move standard generator tests to `generators-stdlib`
