+++ generators-stdlib/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/stdlib/BooleanGeneratorTest.kt
@@ -1,6 +1,6 @@
-package com.github.jcornaz.kwik.generator
+package com.github.jcornaz.kwik.generator.stdlib
 
-import com.github.jcornaz.kwik.generator.stdlib.booleans
+import com.github.jcornaz.kwik.generator.Generator
 import com.github.jcornaz.kwik.generator.test.AbstractGeneratorTest
 import kotlin.test.Test
 import kotlin.test.assertTrue
