+++ generators-stdlib/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/stdlib/FiniteValueSetGenerator.kt
@@ -1,20 +1,22 @@
-package com.github.jcornaz.kwik.generator
+package com.github.jcornaz.kwik.generator.stdlib
 
-import com.github.jcornaz.kwik.generator.stdlib.enum
+import com.github.jcornaz.kwik.generator.Generator
 import com.github.jcornaz.kwik.generator.test.AbstractGeneratorTest
 import kotlin.test.Test
 import kotlin.test.assertFailsWith
 import kotlin.test.assertTrue
 
-class EnumGeneratorTest : AbstractGeneratorTest() {
-    enum class ExampleEnum { A, B, C, D, E }
-    enum class EmptyEnum
-
-    override val generator: Generator<ExampleEnum> = Generator.enum()
+class FiniteValueSetGenerator : AbstractGeneratorTest() {
+    override val generator: Generator<Int> = Generator.of(0, 1, 2, 3, 4)
 
     @Test
     fun failsIfNoGivenSample() {
-        assertFailsWith<IllegalArgumentException> { Generator.enum<EmptyEnum>() }
+        assertFailsWith<IllegalArgumentException> { Generator.of<Int>() }
+    }
+
+    @Test
+    fun generateOnlyGivenSamples() {
+        assertTrue(generator.randoms(-78).take(200).all { it in 0..4 })
     }
 
     @Test
@@ -22,7 +24,7 @@ class EnumGeneratorTest : AbstractGeneratorTest() {
         val counts = IntArray(5)
 
         generator.randoms(-37).take(10_000).forEach {
-            counts[it.ordinal]++
+            counts[it]++
         }
 
         assertTrue(counts.all { it in 1800..2200 })
