+++ generators-stdlib/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/stdlib/FloatGeneratorTest.kt
@@ -1,9 +1,6 @@
-package com.github.jcornaz.kwik.generator
+package com.github.jcornaz.kwik.generator.stdlib
 
-import com.github.jcornaz.kwik.generator.stdlib.floats
-import com.github.jcornaz.kwik.generator.stdlib.negativeFloats
-import com.github.jcornaz.kwik.generator.stdlib.nonZeroFloats
-import com.github.jcornaz.kwik.generator.stdlib.positiveFloats
+import com.github.jcornaz.kwik.generator.Generator
 import com.github.jcornaz.kwik.generator.test.AbstractGeneratorTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
